utils: rename tests.utils.htmx.test.py file into testing.py
Avoid any misunderstanding with tests.utils.htmx.tests.py containing true tests
