+++ tests/gps/test_views.py
@@ -32,7 +32,7 @@ from tests.users.factories import (
     LaborInspectorFactory,
     PrescriberFactory,
 )
-from tests.utils.htmx.test import assertSoupEqual, update_page_with_htmx
+from tests.utils.htmx.testing import assertSoupEqual, update_page_with_htmx
 from tests.utils.testing import assertSnapshotQueries, get_session_name, parse_response_to_soup, pretty_indented
