+++ tests/utils/htmx/tests.py
@@ -5,7 +5,7 @@ from bs4 import BeautifulSoup
 from django.urls import reverse
 from django_htmx.middleware import HtmxDetails
 
-from tests.utils.htmx.test import assertSoupEqual, update_page_with_htmx
+from tests.utils.htmx.testing import assertSoupEqual, update_page_with_htmx
 from tests.utils.testing import parse_response_to_soup
