+++ tests/www/apply/test_list_for_job_seeker.py
@@ -20,7 +20,7 @@ from tests.job_applications.factories import (
 )
 from tests.prescribers.factories import PrescriberOrganizationWithMembershipFactory
 from tests.users.factories import JobSeekerFactory
-from tests.utils.htmx.test import assertSoupEqual, update_page_with_htmx
+from tests.utils.htmx.testing import assertSoupEqual, update_page_with_htmx
 from tests.utils.testing import assertSnapshotQueries, parse_response_to_soup, pretty_indented
