+++ tests/www/apply/test_list_for_siae.py
@@ -25,7 +25,7 @@ from tests.job_applications.factories import JobApplicationFactory, JobApplicati
 from tests.jobs.factories import create_test_romes_and_appellations
 from tests.prescribers.factories import PrescriberOrganizationWithMembershipFactory
 from tests.users.factories import JobSeekerFactory
-from tests.utils.htmx.test import assertSoupEqual, update_page_with_htmx
+from tests.utils.htmx.testing import assertSoupEqual, update_page_with_htmx
 from tests.utils.testing import (
     assert_previous_step,
     assertSnapshotQueries,
