+++ tests/www/employee_record_views/test_list.py
@@ -27,7 +27,7 @@ from tests.job_applications.factories import (
     JobApplicationFactory,
     JobApplicationWithCompleteJobSeekerProfileFactory,
 )
-from tests.utils.htmx.test import assertSoupEqual, update_page_with_htmx
+from tests.utils.htmx.testing import assertSoupEqual, update_page_with_htmx
 from tests.utils.testing import assertSnapshotQueries, parse_response_to_soup, pretty_indented
