+++ tests/www/geiq_assessments_views/test_views_for_geiq.py
@@ -22,7 +22,7 @@ from tests.geiq_assessments.factories import (
 )
 from tests.institutions.factories import InstitutionFactory, InstitutionMembershipFactory
 from tests.users.factories import EmployerFactory, JobSeekerFactory, LaborInspectorFactory, PrescriberFactory
-from tests.utils.htmx.test import assertSoupEqual, update_page_with_htmx
+from tests.utils.htmx.testing import assertSoupEqual, update_page_with_htmx
 from tests.utils.testing import (
     assertSnapshotQueries,
     default_storage_ls_files,
