+++ tests/www/job_seekers_views/test_list.py
@@ -30,7 +30,7 @@ from tests.users.factories import (
     LaborInspectorFactory,
     PrescriberFactory,
 )
-from tests.utils.htmx.test import assertSoupEqual, update_page_with_htmx
+from tests.utils.htmx.testing import assertSoupEqual, update_page_with_htmx
 from tests.utils.testing import assertSnapshotQueries, parse_response_to_soup, pretty_indented
 from tests.www.apply.test_submit import fake_session_initialization
