+++ tests/www/search/tests.py
@@ -24,7 +24,7 @@ from tests.job_applications.factories import JobApplicationFactory
 from tests.jobs.factories import create_test_romes_and_appellations
 from tests.prescribers.factories import PrescriberOrganizationFactory
 from tests.users.factories import PrescriberFactory
-from tests.utils.htmx.test import assertSoupEqual, update_page_with_htmx
+from tests.utils.htmx.testing import assertSoupEqual, update_page_with_htmx
 from tests.utils.testing import assertSnapshotQueries, parse_response_to_soup
